+++ src/app.ts
@@ -5,7 +5,9 @@ import YAML from 'yamljs';
 import userRouter from './resources/users/user.router';
 import boardRouter from './resources/boards/board.router';
 import taskRouter from './resources/tasks/task.router';
-import { errorLogger, requestLogger, uncaughtExceptionLogger, unhandledRejectionLogger } from './middleware/loggers';
+import errorLogger from './middleware/error-logger';
+import requestLogger from './middleware/request-logger';
+import uncaughtErrorHandler from './error-handlers/uncaught-error-handler';
 
 const app = express();
 const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
@@ -33,13 +35,10 @@ app.use('/boards/:boardId/tasks', taskRouter);
 app.use(errorLogger);
 
 process.on('uncaughtException', (err: Error) => {
-  uncaughtExceptionLogger(err);
+  uncaughtErrorHandler(err);
   process.exit(1);
 });
 
-process.on('unhandledRejection', (err: Error) => {
-  unhandledRejectionLogger(err);
-  process.exit(2);
-});
+process.on('unhandledRejection', uncaughtErrorHandler);
 
 export default app;
